Enhancement in security and robustness

+++ sources/classes/NedStars/FileSystem.php
@@ -256,13 +256,16 @@ class NedStars_FileSystem {
 	/**
 	 * Delete given dir recursive
 	 *
-	 * @param String $path absoulte path of folder to delete
+	 * @param String $path absolute path of folder to delete
 	 *
 	 * @return Boolean
 	 */
 	static public function deleteDir($path) {
 		// make sure path exists with line ending
 		$path = self::_getValidatedDir($path);
+        if ($path == DIRECTORY_SEPARATOR) {
+            throw new NedStars_FileSystemException('Not allowed to remove '.DIRECTORY_SEPARATOR, NedStars_FileSystemException::ACTION_NOT_ALLOWED);
+        }
 
 		return NedStars_Execution::run('rm -rf '.escapeshellarg($path), true);
 	}
@@ -276,6 +279,9 @@ class NedStars_FileSystem {
 	static public function deleteDirContent($path) {
 		// make sure path exists with line ending
 		$path = self::_getValidatedDir($path);
+        if ($path == DIRECTORY_SEPARATOR) {
+            throw new NedStars_FileSystemException('Not allowed to remove '.DIRECTORY_SEPARATOR, NedStars_FileSystemException::ACTION_NOT_ALLOWED);
+        }
 
 		return NedStars_Execution::run('rm -rf '.escapeshellarg($path.DIRECTORY_SEPARATOR), true);
 	}
